Added default keyboardSize #995
Currently keyboard size isn't stored, we need to set an initial value ("XS", "S", "M" or "L"). Legacy uses "M" equivalent as initial value. We need to adjust webos-keyboard as well to pick this value up and store it in case it changes. This is a partial fix for bug #995